+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## 5.2.0
 
-* Expose option for "fastcollate" to allow proxy for <=5.0.5 processing methods when combined with `-dupmode t`.
+* Expose option for "legacy" to allow for <=5.0.5 processing methods.
 
 ## 5.1.0